rename sh_utils dir -> env_utils
